Re-organize exports of `core-consensus` crate
