+++ code/crates/core-consensus/src/handle/proposal.rs
@@ -3,9 +3,8 @@ use crate::handle::signature::verify_signature;
 use crate::handle::validator_set::get_validator_set;
 use crate::input::Input;
 use crate::prelude::*;
-use crate::types::ConsensusMsg;
+use crate::types::{ConsensusMsg, ProposedValue, SignedConsensusMsg};
 use crate::util::pretty::PrettyProposal;
-use crate::{ProposedValue, SignedConsensusMsg};
 
 pub async fn on_proposal<Ctx>(
     co: &Co<Ctx>,
