+++ code/crates/core-consensus/src/state.rs
@@ -4,9 +4,11 @@ use tracing::{debug, warn};
 use malachitebft_core_driver::Driver;
 use malachitebft_core_types::*;
 
+use crate::full_proposal::{FullProposal, FullProposalKeeper};
 use crate::input::Input;
+use crate::params::Params;
+use crate::types::ProposedValue;
 use crate::util::max_queue::MaxQueue;
-use crate::{FullProposal, FullProposalKeeper, Params, ProposedValue};
 
 /// The state maintained by consensus for processing a [`Input`][crate::Input].
 pub struct State<Ctx>
