+++ code/examples/channel/src/state.rs
@@ -10,11 +10,10 @@ use rand::{Rng, SeedableRng};
 use sha3::Digest;
 use tracing::{debug, error};
 
-use malachitebft_app_channel::app::consensus::ProposedValue;
 use malachitebft_app_channel::app::streaming::{StreamContent, StreamMessage};
 use malachitebft_app_channel::app::types::codec::Codec;
 use malachitebft_app_channel::app::types::core::{CommitCertificate, Round, Validity};
-use malachitebft_app_channel::app::types::{LocallyProposedValue, PeerId};
+use malachitebft_app_channel::app::types::{LocallyProposedValue, PeerId, ProposedValue};
 use malachitebft_test::codec::proto::ProtobufCodec;
 use malachitebft_test::{
     Address, Genesis, Height, ProposalData, ProposalFin, ProposalInit, ProposalPart, TestContext,
