+++ code/examples/channel/src/streaming.rs
@@ -1,9 +1,9 @@
 use std::cmp::Ordering;
 use std::collections::{BTreeMap, BinaryHeap, HashSet};
 
-use malachitebft_app_channel::app::consensus::PeerId;
 use malachitebft_app_channel::app::streaming::{Sequence, StreamId, StreamMessage};
 use malachitebft_app_channel::app::types::core::Round;
+use malachitebft_app_channel::app::types::PeerId;
 use malachitebft_test::{Address, Height, ProposalInit, ProposalPart};
 
 struct MinSeq<T>(StreamMessage<T>);
